% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(github_actions, host_name, verbose)
}
\arguments{
\item{github_actions}{the GitHub Actions environment value,
as obtained by using \code{Sys.getenv("GITHUB_ACTIONS")}}

\item{host_name}{the name of the host,
as obtained by using \code{Sys.getenv("HOSTNAME")}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
