% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_login_text.R
\name{get_login_text}
\alias{get_login_text}
\title{Get the Uppmax login text}
\usage{
get_login_text()
}
\value{
the Uppmax login text
}
\description{
Get the Uppmax login text
}
\examples{
cat(get_login_text())
}
\author{
Richèl J.C. Bilderbeek
}
