% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_where_i_am.R
\name{get_where_i_am}
\alias{get_where_i_am}
\title{Describe in which environment this script is run}
\usage{
get_where_i_am(
  host_name = Sys.getenv("HOSTNAME"),
  github_actions = Sys.getenv("GITHUB_ACTIONS")
)
}
\arguments{
\item{host_name}{the name of the host,
as obtained by using \code{Sys.getenv("HOSTNAME")}}

\item{github_actions}{the GitHub Actions environment value,
as obtained by using \code{Sys.getenv("GITHUB_ACTIONS")}}
}
\value{
\tabular{ll}{
   Environment \tab Return value \cr
   Bianca login node \tab \code{bianca_login_node} \cr
   Bianca runner node \tab \code{bianca_runner_node} \cr
   Rackham login node \tab \code{rackham_login_node} \cr
   Rackham runner ode \tab \code{rackham_runner_node} \cr
   GitHub Actions \tab \code{github_actions} \cr
   Else \tab \code{unknown} \cr
}
}
\description{
Describe in which environment this script is run
}
\examples{
get_where_i_am()
}
\author{
Richèl J.C. Bilderbeek
}
