% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_gha.R
\name{is_on_gha}
\alias{is_on_gha}
\title{Determine if the environment is GitHub Actions}
\usage{
is_on_gha(github_actions = Sys.getenv("GITHUB_ACTIONS"))
}
\arguments{
\item{github_actions}{the GitHub Actions environment value,
as obtained by using \code{Sys.getenv("GITHUB_ACTIONS")}}
}
\description{
Determine if the environment is GitHub Actions
}
\examples{
is_on_gha()
}
\author{
Richèl J.C. Bilderbeek
}
