% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_rackham_runner_node.R
\name{is_on_rackham_runner_node}
\alias{is_on_rackham_runner_node}
\title{Determine if the environment is a Rackham runner node}
\usage{
is_on_rackham_runner_node(host_name = Sys.getenv("HOSTNAME"))
}
\arguments{
\item{host_name}{the name of the host,
as obtained by using \code{Sys.getenv("HOSTNAME")}}
}
\description{
Determine if the environment is a Rackham runner node
}
\examples{
is_on_rackham_runner_node()
is_on_rackham_runner_node(host_name = "r123")
}
\seealso{
Use \link{is_on_rackham_login_node} to determine
if the environment is a Rackham login node.

Use \link{is_on_bianca_runner_node} to determine
if the environment is a Bianca runner node.
}
\author{
Richèl J.C. Bilderbeek
}
